requirements.txt
huggingface_hub==0.15.1
langchain==0.0.353
langchain_community==0.0.7
langchain_core==0.1.4
numpy==1.24.3
streamlit==1.29.0
sentence-transformers
InstructorEmbedding
faiss-cpu
